% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admis.R
\name{admis}
\alias{admis}
\title{Initialization of admissibility test recursion.}
\usage{
admis(nvec, zvec)
}
\arguments{
\item{nvec}{The vector of numbers of observations per cluster.}

\item{zvec}{The vector of observed z-values (the sum of successses in cluster)}
}
\description{
This function initializes the admissability test recursion for exact clustered logistic regression.  The matrix of admissable vectors starts as observed values, zvec.  The final output is a matrix whose columns are z-vectors that give the same sufficient statstics as those observed.
}
\examples{
admis()
}
\keyword{Clustered}
\keyword{Exact}
\keyword{Logistic}
\keyword{Regression}
