% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admis2.R
\name{admis2}
\alias{admis2}
\title{Initialization of admissibility test recursion, two-level model.}
\usage{
admis2(nvec, zvec, jvec)
}
\arguments{
\item{nvec}{The vector of numbers of observations per cluster.}

\item{zvec}{The vector of observed z-values (the sum of successses in cluster).}

\item{jvec}{The numbers of inner clusters in each outer cluster.}
}
\description{
This function initializes the admissability test recursion for exact clustered logistic regression, two-level model.  The matrix of admissable vectors starts as observed values, zvec.  The final output is a matrix whose columns are z-vectors that give the same sufficient statstics as those observed.
}
\examples{
admis2()
}
\keyword{Clustered}
\keyword{Exact}
\keyword{Logistic}
\keyword{Regression}
