% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admisrec.R
\name{admisrec}
\alias{admisrec}
\title{Admissibility test recursion.}
\usage{
admisrec(nvec, zmat, zvec, s1, s2)
}
\arguments{
\item{nvec}{The vector of numbers of observations per cluster.}

\item{zmat}{The vector of observed z-values.}

\item{zvec}{The vector of (possibly incomplete) z-values.}

\item{s1}{Observed value of the first sufficient statistic.}

\item{s2}{Observed value of the second sufficient statistic.}
}
\description{
This function performs the admissability test recursion for exact clustered logistic regression.  The matrix of admissable vectors starts as observed values, zvec.  The final output is a matrix whose columns are z-vectors that give the same sufficient statstics as those observed.
}
\examples{
admisrec()
}
\keyword{Clustered}
\keyword{Exact}
\keyword{Logistic}
\keyword{Regression}
