% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admisrec2.R
\name{admisrec2}
\alias{admisrec2}
\title{Admissibility test recursion, two-level model.}
\usage{
admisrec2(nvec, zmat, zvec, jvec, s1, s2, s3)
}
\arguments{
\item{nvec}{The vector of numbers of observations per cluster.}

\item{zmat}{The vector of observed z-values.}

\item{zvec}{The vector of (possibly incomplete) z-values.}

\item{jvec}{The numbers of inner clusters in each outer cluster.}

\item{s1}{Observed value of the first sufficient statistic.}

\item{s2}{Observed value of the second sufficient statistic.}

\item{s2}{Observed value of the third sufficient statistic.}
}
\description{
This function performs the admissability test recursion for exact clustered logistic regression, two-level model.  The matrix of admissable vectors starts as observed values, zvec.  The final output is a matrix whose columns are z-vectors that give the same sufficient statstics as those observed.
}
\examples{
admisrec2()
}
\keyword{Clustered}
\keyword{Exact}
\keyword{Logistic}
\keyword{Regression}
