% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprod.R
\name{comprod}
\alias{comprod}
\title{Calculation of the product of combinations used in conditional probability distribution for exact clustered logistic regression.}
\usage{
comprod(nvec, zvec)
}
\arguments{
\item{nvec}{The vector of numbers of observations per cluster.}

\item{zvec}{The vector of observed z-values (the sum of successses in cluster)}
}
\description{
This function calculates the product of combinations used in conditional probability distribution for exact clustered logistic regression.
}
\examples{
comprod()
}
\keyword{Clustered}
\keyword{Exact}
\keyword{Logistic}
\keyword{Regression}
