% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremez.R
\name{extremez}
\alias{extremez}
\title{Calculation of all admissable z-vectors with corresponding t-statistics at least as large as the observed value of t.}
\usage{
extremez(zadmis, nvec, z, x)
}
\arguments{
\item{zadmis}{The matrix of admissable vectors.}

\item{nvec}{The vector of numbers of observations per cluster.}

\item{z}{The vector of observed z-values.}

\item{x}{A vector of predictor values for each cluster.}
}
\description{
This function performs the calculation of all admissable z-vectors with corresponding t-statistics at least as large as the observed value of t for exact clustered logistic regression.
}
\examples{
extremez()
}
\keyword{Clustered}
\keyword{Exact}
\keyword{Logistic}
\keyword{Regression}
