% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s1calc.R
\name{s1calc}
\alias{s1calc}
\title{Calculation of first sufficient statistic for exact clustered logistic regression.}
\usage{
s1calc(zvec)
}
\arguments{
\item{zvec}{The vector of observed z-values (the sum of successses in cluster)}
}
\description{
This function calculates the first sufficient statistic, s1, for exact clustered logistic regression.
}
\examples{
s1calc()
}
\keyword{Clustered}
\keyword{Exact}
\keyword{Logistic}
\keyword{Regression}
