% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2calc.R
\name{s2calc}
\alias{s2calc}
\title{Calculation of second sufficient statistic for exact clustered logistic regression.}
\usage{
s2calc(zvec, nvec)
}
\arguments{
\item{zvec}{The vector of observed z-values (the sum of successses in cluster)}

\item{nvec}{The vector of numbers of observations per cluster}
}
\description{
This function calculates the second sufficient statistic, s2, for exact clustered logistic regression.
}
\examples{
s2calc()
}
\keyword{Clustered}
\keyword{Exact}
\keyword{Logistic}
\keyword{Regression}
