% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3calc.R
\name{s3calc}
\alias{s3calc}
\title{Calculation of third sufficient statistic for exact clustered logistic regression.}
\usage{
s3calc(zvec, nvec, jvec)
}
\arguments{
\item{zvec}{The vector of observed z-values (the sum of successses in cluster)}

\item{nvec}{The vector of numbers of observations per cluster}

\item{jvec}{The vector of numbers of inner clusters per outer cluster}
}
\description{
This function calculates the third sufficient statistic, s3, for exact clustered logistic regression.
}
\examples{
s3calc()
}
\keyword{Clustered}
\keyword{Exact}
\keyword{Logistic}
\keyword{Regression}
