% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcalc.R
\name{tcalc}
\alias{tcalc}
\title{Calculation of a sufficient statistic for a parameter of iterest for exact clustered logistic regression.}
\usage{
tcalc(zvec, x)
}
\arguments{
\item{zvec}{The vector of observed z-values (the sum of successses in cluster)}

\item{x}{A vector of predictor values for each cluster.}
}
\description{
This function calculates a sufficient statistic, t, for a parameter of iterest for exact clustered logistic regression.
}
\examples{
tcalc()
}
\keyword{Clustered}
\keyword{Exact}
\keyword{Logistic}
\keyword{Regression}
